# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SETU-DRR API",
    "description": "Hazard Red Zone Identification, Carrying Capacity Assessment & Relocation Decision Support Platform.",
    "version": "v1.0.0"
  },
  "paths": {
...
    "/habitations": {
      "get": {
        "tags": [
          "Habitations & Triage"
        ],
        "summary": "Get prioritized habitation triage queue",
        "description": "Returns a paginated list of habitations ranked by Per-Capita Urgency or Caseload. Supports filtering by Administrative Boundary (LGD Code) and 4-tier Triage Category.",
        "operationId": "get_habitations_habitations_get",
        "parameters": [
          {
            "name": "admin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Administrative Unit ID or LGD Code (e.g. 555 for Wayanad)",
              "title": "Admin"
            },
            "description": "Filter by Administrative Unit ID or LGD Code (e.g. 555 for Wayanad)"
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Tier"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Triage Tier (Immediate, Short-term, Medium-term, Mitigate in situ)",
              "title": "Tier"
            },
            "description": "Filter by Triage Tier (Immediate, Short-term, Medium-term, Mitigate in situ)"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortMode",
              "description": "Ranking mode: 'urgency' (PS_j) or 'caseload' (PS_j * population)",
              "default": "urgency"
            },
            "description": "Ranking mode: 'urgency' (PS_j) or 'caseload' (PS_j * population)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Number of records per page (max 200).",
              "default": 50,
              "title": "Limit"
            },
            "description": "Number of records per page (max 200)."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_HabitationListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/habitations/{id}/risk": {
      "get": {
        "tags": [
          "Habitations & Triage"
        ],
        "summary": "Get complete risk dossier for a habitation",
        "description": "Retrieves vulnerability breakdown (SoVI), time-decayed loss history, and priority score explanation.",
        "operationId": "get_habitation_risk_dossier_habitations__id__risk_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Habitation ID (integer primary key).",
              "examples": [
                1
              ],
              "title": "Id"
            },
            "description": "Habitation ID (integer primary key)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitationRiskDossier"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/habitations/{id}/sites": {
      "get": {
        "tags": [
          "Habitations & Triage"
        ],
        "summary": "Get ranked candidate relocation sites for a habitation",
        "description": "Retrieves candidate relocation sites within search radius (default 15 km) ranked by suitability, carrying capacity, and distance. Includes full resource capacity breakdowns and binding constraints.",
        "operationId": "get_habitation_candidate_sites_habitations__id__sites_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Habitation ID (integer primary key).",
              "examples": [
                1
              ],
              "title": "Id"
            },
            "description": "Habitation ID (integer primary key)."
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100.0,
                  "exclusiveMinimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Configurable search radius around habitation in kilometers (default 15 km).",
              "title": "Radius Km"
            },
            "description": "Configurable search radius around habitation in kilometers (default 15 km)."
          },
          {
            "name": "min_suitability",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by minimum composite suitability score (0-100).",
              "title": "Min Suitability"
            },
            "description": "Filter by minimum composite suitability score (0-100)."
          },
          {
            "name": "include_screening",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Whether to include unverified screening-grade candidate sites (default False).",
              "default": false,
              "title": "Include Screening"
            },
            "description": "Whether to include unverified screening-grade candidate sites (default False)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Number of records per page (max 200).",
              "default": 50,
              "title": "Limit"
            },
            "description": "Number of records per page (max 200)."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_CandidateSiteItem_"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Standardized machine-readable error code string."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation of the error."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Unique request tracing identifier."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "Contextual error details or validation breakdown."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Standardized error details conforming to SETU-DRR API error envelope."
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "Error details payload."
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Canonical API error response envelope."
      },
...
      "FeatureContributionDTO": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature",
            "description": "Feature name, e.g. slope_deg, hand_m, dist_to_road_m, rainfall_72h."
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Observed or calculated feature value."
          },
          "contribution": {
            "type": "number",
            "title": "Contribution",
            "description": "Attribution weight or SHAP value contribution."
          },
          "method": {
            "type": "string",
            "title": "Method",
            "description": "Explanation method ('treeshap', 'kernelshap', 'heuristic', 'pca_loading').",
            "default": "heuristic"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank",
            "description": "Rank of this feature in importance (1 = top contributing factor)."
          }
        },
        "type": "object",
        "required": [
          "feature",
          "value",
          "contribution"
        ],
        "title": "FeatureContributionDTO",
        "description": "Single feature attribution item for a cell or habitation."
      },
...
      "HabitationListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lgd_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lgd Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "village"
          },
          "admin_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Id"
          },
          "admin_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Name"
          },
          "population": {
            "type": "integer",
            "title": "Population",
            "default": 0
          },
          "households": {
            "type": "integer",
            "title": "Households",
            "default": 0
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score",
            "description": "Per-capita urgency score PS_j."
          },
          "caseload_score": {
            "type": "number",
            "title": "Caseload Score",
            "description": "Caseload urgency score PS_j * population."
          },
          "tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Tier"
              },
              {
                "type": "null"
              }
            ],
            "description": "Four-tier triage category (None if unclassified/monitoring)."
          },
          "prz_overlap_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Prz Overlap Pct",
            "description": "Percentage of built area inside PRZ."
          },
          "dominant_hazard": {
            "type": "string",
            "title": "Dominant Hazard",
            "default": "landslide"
          },
          "centroid": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Centroid",
            "description": "[longitude, latitude]"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "default": "baseline-v1"
          },
          "scoring_version": {
            "type": "string",
            "title": "Scoring Version",
            "default": "priority-v1.0"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version",
            "default": "v1.0"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "priority_score",
          "caseload_score",
          "prz_overlap_pct",
          "centroid"
        ],
        "title": "HabitationListItem",
        "description": "Item in prioritized habitation triage queue (GET /habitations)."
      },
      "HabitationRiskDossier": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lgd_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lgd Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "village"
          },
          "admin_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Id"
          },
          "admin_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Name"
          },
          "population": {
            "type": "integer",
            "title": "Population"
          },
          "households": {
            "type": "integer",
            "title": "Households"
          },
          "centroid": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Centroid"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          },
          "caseload_score": {
            "type": "number",
            "title": "Caseload Score"
          },
          "tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Tier"
              },
              {
                "type": "null"
              }
            ],
            "description": "Four-tier triage category (None if unclassified/monitoring)."
          },
          "triage_rationale": {
            "type": "string",
            "title": "Triage Rationale"
          },
          "prz_overlap_pct": {
            "type": "number",
            "title": "Prz Overlap Pct"
          },
          "hazard_intensity": {
            "type": "number",
            "title": "Hazard Intensity"
          },
          "decayed_loss_score": {
            "type": "number",
            "title": "Decayed Loss Score"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "default": "baseline-v1"
          },
          "scoring_version": {
            "type": "string",
            "title": "Scoring Version",
            "default": "priority-v1.0"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version",
            "default": "v1.0"
          },
          "data_quality": {
            "type": "string",
            "title": "Data Quality",
            "default": "synthetic"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 1.0
          },
          "calculated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calculated At"
          },
          "vulnerability": {
            "$ref": "#/components/schemas/VulnerabilityBreakdownDTO"
          },
          "past_disasters": {
            "items": {
              "$ref": "#/components/schemas/LossEventDTO"
            },
            "type": "array",
            "title": "Past Disasters"
          },
          "top_contributing_factors": {
            "items": {
              "$ref": "#/components/schemas/FeatureContributionDTO"
            },
            "type": "array",
            "title": "Top Contributing Factors"
          },
          "screening_grade": {
            "type": "string",
            "title": "Screening Grade",
            "default": "Screening Grade: Cell-level screening and prioritisation tool. Geotechnical investigation, hydraulic study, and community consultation required before executing relocation orders."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "population",
          "households",
          "centroid",
          "priority_score",
          "caseload_score",
          "triage_rationale",
          "prz_overlap_pct",
          "hazard_intensity",
          "decayed_loss_score",
          "vulnerability"
        ],
        "title": "HabitationRiskDossier",
        "description": "Full risk dossier for a single habitation (GET /habitations/{id}/risk)."
      },
...
      "LossEventDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ts": {
            "type": "string",
            "format": "date",
            "title": "Ts"
          },
          "hazard_type": {
            "type": "string",
            "title": "Hazard Type"
          },
          "fatalities": {
            "type": "integer",
            "title": "Fatalities",
            "default": 0
          },
          "injured": {
            "type": "integer",
            "title": "Injured",
            "default": 0
          },
          "houses_damaged": {
            "type": "integer",
            "title": "Houses Damaged",
            "default": 0
          },
          "severity": {
            "type": "number",
            "title": "Severity",
            "default": 1.0
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ts",
          "hazard_type",
          "source"
        ],
        "title": "LossEventDTO"
      },
      "PaginatedResponse_CandidateSiteItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandidateSiteItem"
            },
            "type": "array",
            "title": "Items",
            "description": "Page of records."
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total matching records count."
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "description": "Page limit."
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Page offset."
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether subsequent records exist."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PaginatedResponse[CandidateSiteItem]"
      },
      "PaginatedResponse_HabitationListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HabitationListItem"
            },
            "type": "array",
            "title": "Items",
            "description": "Page of records."
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total matching records count."
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "description": "Page limit."
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Page offset."
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether subsequent records exist."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PaginatedResponse[HabitationListItem]"
      },
...
      "SortMode": {
        "type": "string",
        "enum": [
          "urgency",
          "caseload"
        ],
        "title": "SortMode",
        "description": "Sort modes for prioritized habitation queues (PRD \u00a76.6, FR-6.3)."
      },
...
      "Tier": {
        "type": "string",
        "enum": [
          "immediate",
          "short_term",
          "medium_term",
          "mitigate_in_situ"
        ],
        "title": "Tier",
        "description": "Four-tier triage categorization for relocation planning (PRD \u00a76.7)."
      },
...
      "VulnerabilityBreakdownDTO": {
        "properties": {
          "v_demographic": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "V Demographic"
          },
          "v_structural": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "V Structural"
          },
          "v_access": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "V Access"
          },
          "v_economic": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "V Economic"
          },
          "v_index": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "V Index"
          },
          "is_district_flat": {
            "type": "boolean",
            "title": "Is District Flat",
            "default": false
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Downscaling validation & PCA weights metadata"
          }
        },
        "type": "object",
        "required": [
          "v_demographic",
          "v_structural",
          "v_access",
          "v_economic",
          "v_index"
        ],
        "title": "VulnerabilityBreakdownDTO"
      },
...
    }
  }
}